BigDecimal

+++ src/com/company/testmoney/NegativeAmountException.java
@@ -9,7 +9,7 @@ public class NegativeAmountException extends RuntimeException {
         this.amount = amount;
     }
 
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
